fix client side tests
