+++ client/src/components/Jobs/Jobs.test.js
@@ -8,6 +8,7 @@ import '@testing-library/jest-dom/extend-expect';
 import { BrowserRouter } from 'react-router-dom';
 import { initialState as notifications } from '../../redux/reducers/notifications.reducers';
 import { initialState as workbench } from '../../redux/reducers/workbench.reducers';
+import { initialState as jobs } from '../../redux/reducers/jobs.reducers';
 
 const mockStore = configureStore();
 const initialMockState = {
@@ -31,6 +32,29 @@ function renderJobsComponent(store) {
 }
 
 describe('Jobs View', () => {
+  it('should dispatch the get jobs event', () => {
+    const store = mockStore({
+      notifications,
+      jobs: { ...jobs, list: jobsList },
+      workbench: { ...workbench, config: { hideDataFiles: false } },
+      apps: {
+        appIcons: {},
+      },
+    });
+
+    render(
+      <Provider store={store}>
+        <BrowserRouter>
+          <Jobs />
+        </BrowserRouter>
+      </Provider>
+    );
+
+    expect(store.getActions()).toEqual([
+      { type: 'GET_JOBS', params: { offset: 0, queryString: '' } },
+    ]);
+  });
+
   it('renders jobs', () => {
     const store = mockStore({
       jobs: {
@@ -60,6 +84,7 @@ describe('Jobs View', () => {
       apps: {
         ...appIconMockState,
       },
+      notifications,
       workbench: {
         ...workbench,
         config: { hideDataFiles: false },
